issue #192 refactor: split shared utils in ScheduleModal to calendarUtils
